+++ src/lib.rs
@@ -122,7 +122,8 @@ pub mod types {
     };
     pub use super::error::types::{ErrorKind, ErrorResponse};
     pub use super::receiving::types::{
-        InboundAttachment, InboundAttatchmentId, InboundEmail, InboundEmailId,
+        ForwardInboundEmailResponse, ForwardReceivingEmail, InboundAttachment,
+        InboundAttatchmentId, InboundEmail, InboundEmailId,
     };
     pub use super::segments::types::{CreateSegmentResponse, Segment, SegmentId};
     pub use super::templates::types::{